hf-space-deployment/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
numpy>=1.24.0
gradio>=4.0.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
